SDK: original StApi class removed, tours method moved to new StApi class, internal ToursResponse
