+++ sdk/src/main/java/com/sygic/travel/sdk/api/StApi.kt
@@ -3,6 +3,7 @@ package com.sygic.travel.sdk.api
 import com.sygic.travel.sdk.api.responseWrappers.MediaResponse
 import com.sygic.travel.sdk.api.responseWrappers.PlaceDetailedResponse
 import com.sygic.travel.sdk.api.responseWrappers.PlacesResponse
+import com.sygic.travel.sdk.api.responseWrappers.TourResponse
 import retrofit2.adapter.rxjava.Result
 import retrofit2.http.GET
 import retrofit2.http.Path
@@ -47,4 +48,12 @@ internal interface StApi {
     fun getPlaceMedia(
             @Path("id") id: String
     ): Observable<Result<MediaResponse>>
+
+    @GET("tours/viator")
+    fun getTours(
+            @Query("destination_id") destinationId: String?,
+            @Query("page") page: Int?,
+            @Query("sort_by") sortBy: String?,
+            @Query("sort_direction") sortDirection: String?
+    ): Observable<Result<TourResponse>>
 }
